replaced buttons below with bottom navigation view, standardised displays
