+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.trackablehabit">
+    package="com.example.trackablehabit" >
 
     <uses-permission android:name="android.permission.RECEIVE_BOOT_COMPLETED" />
     <uses-permission android:name="android.permission.INTERNET" />
@@ -13,15 +13,12 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/AppTheme">
-        <activity
-            android:name=".Rewards"
-            android:parentActivityName=".MainActivity"/>
+        android:theme="@style/AppTheme" >
         <activity android:name=".RegisterActivity" />
-        <activity
-            android:name=".LoginActivity">
+        <activity android:name=".LoginActivity" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
@@ -29,15 +26,9 @@
             android:name=".EditHabit"
             android:parentActivityName=".MainActivity" />
         <activity
-            android:name=".Statistics"
-            android:parentActivityName=".MainActivity" />
-        <activity
             android:name=".AddHabit"
             android:parentActivityName=".MainActivity" />
-        <activity
-            android:name=".UserManual"
-            android:parentActivityName=".MainActivity" />
-        <activity android:name=".MainActivity">
+        <activity android:name=".MainActivity" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -48,7 +39,7 @@
             android:name=".AddReminder"
             android:label="@string/add_reminder"
             android:parentActivityName=".MainActivity"
-            android:theme="@style/AppTheme">
+            android:theme="@style/AppTheme" >
 
             <!-- Parent activity meta-data to support 4.0 and lower -->
             <meta-data
