--- app/src/main/java/com/example/trackablehabit/UserManual.java
@@ -1,23 +1,0 @@
-package com.example.trackablehabit;
-
-import androidx.appcompat.app.ActionBar;
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-
-public class UserManual extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_user_manual);
-
-        ActionBar ab = getSupportActionBar();
-        if (ab != null) {
-            ab.setTitle("User Manual");
-        }
-    }
-}
