+++ app/src/main/res/layout/habit_individual.xml
@@ -29,7 +29,7 @@
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:layout_weight="9"
-            android:background="#A6DFDA"
+            android:background="@color/colorAccent"
             android:orientation="vertical">
 
             <TextView
