+++ app/src/main/res/menu/bottom_navigation.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:id="@+id/nav_home"
+        android:icon="@drawable/ic_home"
+        android:title="@string/habits" />
+
+    <item
+        android:id="@+id/nav_statistics"
+        android:icon="@drawable/ic_stats"
+        android:title="@string/statistics" />
+
+    <item
+        android:id="@+id/nav_rewards"
+        android:icon="@drawable/ic_star_circle"
+        android:title="@string/rewards" />
+
+    <item
+        android:id="@+id/nav_news"
+        android:icon="@drawable/ic_dashboard"
+        android:title="@string/news" />
+
+</menu>
